feat: restructure folder organization into architecture layers
