+++ pubspec.yaml
@@ -7,6 +7,7 @@ environment:
   sdk: ^3.8.1
 
 dependencies:
+  system_command_runner: ^0.1.0
   equatable: ^2.0.7
   flutter:
     sdk: flutter
